Move review pages to a subfolder
We’re preparing to add a new task to this section, so it makes sense
(and is more consistent) to have the existing task in a subfolder.
