+++ server/form-pages/assess/reviewApplication/reviewApplicationAndDocuments/index.ts
@@ -1,0 +1,10 @@
+import { Task } from '../../../utils/decorators'
+
+import Review from './review'
+
+@Task({
+  slug: 'review-application',
+  name: 'Review application and documents',
+  pages: [Review],
+})
+export default class ReviewApplicationAndDocuments {}
